# openapi/unified-openapi.yaml
openapi: 3.1.0
info:
  title: UserManagement
  version: 1.0.0
...
paths:
...
  /api/v1/chat/rooms/{roomId}/messages:
    servers:
      - url: http://localhost:35473
        description: Generated server url
    get:
      tags:
        - Chat
      summary: Get messages from a chat room with cursor pagination
      operationId: getMessages
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
        - name: size
          in: query
          required: false
          schema:
            type: integer
            format: int32
            default: 20
        - name: beforeSeqNum
          in: query
          required: false
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: >-
                  #/components/schemas/ChatManagement_JsendResponseCursorScrollResponseChatMessageResponse
    post:
      tags:
        - Chat
      summary: Send a message to a chat room
      operationId: sendMessage
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChatManagement_SendMessageRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: >-
                  #/components/schemas/ChatManagement_JsendResponseChatMessageResponse
  /api/v1/chat/rooms/{roomId}:
    servers:
      - url: http://localhost:35473
        description: Generated server url
    get:
      tags:
        - Chat
      summary: Get chat room info
      operationId: getRoom
      parameters:
        - name: roomId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: >-
                  #/components/schemas/ChatManagement_JsendResponseChatRoomResponse
components:
  schemas:
...
    ChatManagement_SendMessageRequest:
      type: object
      properties:
        senderName:
          type: string
        content:
          type: string
        replyToSeqNum:
          type: integer
          format: int64
    ChatManagement_ChatMessageResponse:
      type: object
      properties:
        id:
          type: string
        seqNum:
          type: integer
          format: int64
        roomId:
          type: string
        senderId:
          type: string
        senderName:
          type: string
        content:
          type: string
        type:
          type: string
        createdAt:
          type: string
          format: date-time
    ChatManagement_JsendResponseChatMessageResponse:
      type: object
      properties:
        status:
          type: string
        data:
          $ref: '#/components/schemas/ChatManagement_ChatMessageResponse'
        message:
          type: string
    ChatManagement_ChatRoomResponse:
      type: object
      properties:
        roomId:
          type: string
        meetingId:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time
    ChatManagement_JsendResponseChatRoomResponse:
      type: object
      properties:
        status:
          type: string
        data:
          $ref: '#/components/schemas/ChatManagement_ChatRoomResponse'
        message:
          type: string
    ChatManagement_CursorScrollResponseChatMessageResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ChatManagement_ChatMessageResponse'
        size:
          type: integer
          format: int32
        nextPageToken:
          type: string
    ChatManagement_JsendResponseCursorScrollResponseChatMessageResponse:
      type: object
      properties:
        status:
          type: string
        data:
          $ref: >-
            #/components/schemas/ChatManagement_CursorScrollResponseChatMessageResponse
        message:
          type: string
...